fix: close the guard's backtick hole, and finish the profiles rename
Review on #368, all three points actioned.

The `check:convex-skip` guard matched `'skip'` and `"skip"` by substring, so a
backtick spelling walked straight past it. Confirmed by probe before changing
anything — `` `skip` `` was NOT caught — and now matched by one pattern across
all three quote styles, verified again after. `skipped` and `skipper` still do
not trip it.

The same review asked to ignore the word inside comments, which I have
deliberately not done. It is a real false positive — `// never pass 'skip'` in
a db module does fail the build — but that announces itself and costs a
rename, whereas a regex that strips comments can hide a real violation sitting
after a `//` on the same line. Between a loud false positive and a silent
false negative, the guard should keep the loud one. Noted in the script.

`src/shared/profile/` is now `profiles/`, matching `groups`, `rulesets`, `faq`
and the `src/app/db/profiles.ts` this PR renamed. It was the only singular
folder left. Four importers updated.

The doc now says the guard scans `src/app/db/**/*.ts` including `core/`, which
is what the recursive walk actually does.

Verified: root and publisher typecheck, lint, knip, format, check:convex-skip
(six probes: three quote styles, recursion into core/, and two spellings that
must not trip), check:css-orphans, 451 app tests, 279 story tests, and
publisher:release:verify.
