+++ src/app/db/profiles.ts
@@ -1,5 +1,5 @@
-import { profileUserEditFormSchema } from '@shared/profile/validation';
-import type { ProfileUserEditInput } from '@shared/profile/validation';
+import { profileUserEditFormSchema } from '@shared/profiles/validation';
+import type { ProfileUserEditInput } from '@shared/profiles/validation';
 import { useMutation, useQuery } from 'convex/react';
 import type { FunctionReturnType } from 'convex/server';
 import { useEffect, useRef } from 'react';
